fix : fix WithDrawalGroundsType error

+++ src/main/java/com/sosim/server/user/dto/req/UserWithdrawalReq.java
@@ -17,5 +17,5 @@ public class UserWithdrawalReq {
     private LocalDateTime withdrawalDate;
 
     @NotNull
-    private long withdrawalGroundsType;
+    private int withdrawalGroundsType;
 }
